front/syn_metadata: move match-arm pattern classifiers out of front/ast (Slice 2)
Relocates the syn-arm classifier cluster used by match-on-bool /
match-on-integer lowering:
  parse_matches_pat_and_guard, classify_match_bool_exitcases,
  BoolPatternCase, classify_bool_pattern,
  classify_match_switch_exitcases, flatten_or_pattern,
  classify_switch_pattern, default_exitcase, is_default_exitcase.

All helpers operate on syn::Arm / syn::Pat and emit crate::model::ExitCase
without touching the graph builder. front::ast drops the local ExitCase
import (now reached through the helpers).
